IpSpaceMayNotContainWildcard: move to minesweeper
its only user.
